Alter return code based on whether affected or not
It is quite handy to have this program exit with a
non-zero error code if affected (or some other issue)
happens and a zero error code if *likely* not affected.

This changes it so that a boolean is used to track
whether the system is affected (or not) and adjusts
the main() return code based on that boolean.